websocket-service/requirements.txt
flask
flask-socketio
python-dotenv
kafka-python
pyjwt
eventlet